allow queue update

+++ waitingroom/config.go
@@ -9,6 +9,6 @@ type Config struct {
 	EntryDelaySec            int64 // 初回エントリーをDelayさせる秒数
 	QueueEnableSec           int   // 待合室を有効にしておく時間
 	AllowIntervalSec         int   // アクセス許可判定周期
-	AllowUnitNumber          int   // アクセス許可する単位(AllowIntervalSecあたりAllowUnitNumber許可)
+	AllowUnitNumber          int64 // アクセス許可する単位(AllowIntervalSecあたりAllowUnitNumber許可)
 	CacheTTLSec              int   // ローカルメモリキャッシュTTL
 }
